Adding support for more recipe types and forge recipes (#40)
* Foundations for support of different crafting recipe types.

NeuRecipe is now a base class for a recipe which provides common
concepts, such as inputs, outputs and a rendering task.

GuiItemRecipe has been reworked to work with this new NeuRecipe.

NeuManager now parses said recipes. This should be reworked to be a two
step process (first register items, then register recipes).

To keep compatibility with older repo versions, NeuRecipes are parse
lenient and default to a crafting recipe. New recipes should be added in
the `recipes` json field which is an array of json dictionaries, which
have a `type` and other fields depending on the `type` of that recipe.

This also adds support for having multiple recipes for a single item
(e.g. uncrafting storage blocks).

* Remove references in existing code

* Recipe Generation

* ring recipes

* recipe generator v2

* quick forge

* bugfixes and performance improvements

* fix raw craft cost

* reload hotm if you open the hotm tree inv

* add myself to the changelog

* replace quickforge formular with lookup table

* do not crash anymore when opening recipes outside of skyblock

* format coins differently

* remove debug logs

* change recipe generator so that it doesnt crash old versions